Fixes   : Crypto and Storage suites.
Crypto  : Remove the wrong inputs for psa_key_derivation_input_bytes() against
          PSA_ALG_HKDF(PSA_ALG_SHA_256).
Details : Since the focus of the test logic in test_c020 and test_c021 is to verify
          psa_key_derivation_output_bytes() and psa_key_derivation_output_key(),
          tests were checking psa_key_derivation_input_bytes() against hard coded
          PSA_SUCCESS return value.

          As the wrong input check for psa_key_derivation_input_bytes() against
          PSA_ALG_HKDF(PSA_ALG_SHA_256) already verified as part of test_c009,
          the test inputs are removed(PSA_KEY_DERIVATION_INPUT_LABEL and
          PSA_KEY_DERIVATION_INPUT_SEED).

Storage : test_s013 - The argument 'p_data_length' type associated with psa_ps_get
                      and psa_it_get is corrected.

          test_s011 - The error check against psa_ps_create updated.

          test_s002 - Updated test logic to do metadata check for storage element
                      which was created using PSA_STORAGE_FLAG_WRITE_ONCE in the
                      previous run.

          test_s012 - Use psa_storage_uid_t typed values for uid instead of hard
                      coded value.

+++ api-tests/dev_apis/storage/test_s001/test_s001.c
@@ -36,8 +36,8 @@ const client_test_t              s001_storage_test_list[] = {
 
 static int32_t sst_calls_without_set_call(storage_function_code_t fCode, psa_storage_uid_t p_uid)
 {
-    int32_t  status;
-    uint32_t p_data_length = 0;
+    int32_t status;
+    size_t  p_data_length = 0;
 
     /* get() without using set() before */
     val->print(PRINT_TEST, "[Check 1] Call get API for UID %d which is not set\n", p_uid);
@@ -81,8 +81,8 @@ static int32_t sst_set_and_remove(storage_function_code_t fCode, psa_storage_uid
 
 static int32_t sst_calls_after_uid_remove(storage_function_code_t fCode, psa_storage_uid_t p_uid)
 {
-    int32_t  status;
-    uint32_t p_data_length = 0;
+    int32_t status;
+    size_t  p_data_length = 0;
 
     /* get() for UID which is removed */
     val->print(PRINT_TEST, "[Check 4] Call get API for UID %d which is removed\n", p_uid);
@@ -105,8 +105,8 @@ static int32_t sst_calls_after_uid_remove(storage_function_code_t fCode, psa_sto
 
 static int32_t sst_calls_with_different_uid(storage_function_code_t fCode, psa_storage_uid_t p_uid)
 {
-    int32_t  status;
-    uint32_t p_data_length = 0;
+    int32_t status;
+    size_t  p_data_length = 0;
 
     /* set() a UID */
     val->print(PRINT_TEST, "Set storage for UID %d\n", p_uid);
